fix some bugs in Address and edit some DAO functions in Comment

+++ src/tw/dp103g3/address/Address.java
@@ -5,7 +5,7 @@ public class Address {
 	private int mem_id;
 	private String name;
 	private String info;
-	private byte state;
+	private int state;
 	private double latitude;
 	private double longitude;
 	
@@ -32,7 +32,7 @@ public class Address {
 		return address.getLatitude() == this.getLatitude() && address.getLongitude() == this.getLongitude();
 	}
 
-	public Address(int id, int mem_id, String name, String info, byte state, double latitude, double longitude) {
+	public Address(int id, int mem_id, String name, String info, int state, double latitude, double longitude) {
 		this.id = id;
 		this.mem_id = mem_id;
 		this.name = name;
@@ -41,7 +41,7 @@ public class Address {
 		this.latitude = latitude;
 		this.longitude = longitude;
 	}
-
+	
 	public Address(int id, String name, String info, double latitude, double longitude) {
 		super();
 		this.id = id;
@@ -85,11 +85,11 @@ public class Address {
 		this.info = info;
 	}
 
-	public byte getState() {
+	public int getState() {
 		return state;
 	}
 
-	public void setState(byte state) {
+	public void setState(int state) {
 		this.state = state;
 	}
